all use Compute, and many skip slow
