+++ asv_bench/asv.conf.json
@@ -65,6 +65,7 @@
       "xskillscore": [""],
       "dask": [""],
       "cftime": [""],
+      "netcdf4": [""],
     },
 
     // Combinations of libraries/python versions can be excluded/included
