+++ docs/source/contributing.rst
@@ -216,17 +216,18 @@ Preparing Pull Requests
    cause unexpected performance regressions.  You can run specific benchmarks
    using the ``-b`` flag, which takes a regular expression.  For example, this
    will only run tests from a
-   ``asv_bench/benchmarks/benchmarks_HindcastEnsemble.py`` file::
+   ``asv_bench/benchmarks/benchmarks_PredictionEnsemble.py`` file::
 
-      $ asv continuous -f 1.1 upstream/main HEAD -b ^benchmarks_HindcastEnsemble
+      $ asv continuous -f 1.1 upstream/main HEAD -b ^benchmarks_PredictionEnsemble
 
    If you want to only run a specific group of tests from a file, you can do it
    using ``.`` as a separator. For example::
 
-      $ asv continuous -f 1.1 upstream/main HEAD -b benchmarks_HindcastEnsemble.Compute.time_PredictionEnsemble_bootstrap
+      $ asv continuous -f 1.1 upstream/main HEAD -b benchmarks_PredictionEnsemble.GenerateSmallHindcastEnsemble.time_PredictionEnsemble_bootstrap
 
    will only run the ``time_PredictionEnsemble_bootstrap`` benchmark of class
-   ``Compute`` defined in ``benchmarks_HindcastEnsemble.py``.
+   ``GenerateSmallHindcastEnsemble`` defined in ``benchmarks_PredictionEnsemble.py``.
+
 
 #. Create a new changelog entry in `CHANGELOG.rst <CHANGELOG.html>`_:
